Calculate potential mineralizable N for agricultural soils default values are taken from global ISRIC database

Calculate potential mineralizable N for agricultural soils default values are taken from global ISRIC database

Usage

osi_conv_npmn(A_N_RT, A_CLAY_MI, med_PMN = 51.9, med_NRT = 1425, med_CLAY = 26)

Arguments

A_N_RT

(numeric) Soil organic N content (mg N / kg) of top soil

A_CLAY_MI

(numeric) Clay content (%)

med_PMN

(numeric) Regional median value of PMN (mg N / kg)

med_NRT

(numeric) Regional median value of soil organic N content (mg N / kg)

med_CLAY

(numeric) Regional median value of clay content (%)