Estimate soil organic matter and nutrient values (-)

Estimate soil organic matter and nutrient values (-)

Usage

osi_conv_som(
  element,
  A_SOM_LOI = NA_real_,
  A_C_OF = NA_real_,
  A_N_RT = NA_real_,
  A_CN_FR = NA_real_
)

Arguments

element

(character) the method requested to be calculated

A_SOM_LOI

(numeric) The organic matter content of the soil (%)

A_C_OF

(numeric) The organic carbon content of the soil in g C / kg

A_N_RT

(numeric) The organic nitrogen content of the soil in mg N / kg

A_CN_FR

(numeric) The C-to-N ratio of the soil organic matter (-)