Calculate the magnesium availability index in United Kingdom

This function calculates the magnesium availability.

Usage

osi_c_magnesium_uk(B_LU, A_SOM_LOI, A_MG_AN)

Arguments

B_LU

(numeric) The crop code

A_SOM_LOI

(numeric) The percentage organic matter in the soil

A_MG_AN

(numeric) The Mg-content of the soil extracted with ammonium nitrate (mg Mg /kg)

Value

The magnesium availability index in United Kingdom derived from extractable soil Mg fractions. A numeric value.

Examples

osi_c_magnesium_uk(B_LU = 265,A_SOM_LOI=3,A_MG_AN = 50)
#>    [1] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>    [8] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [15] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [22] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [29] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [36] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [43] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [50] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [57] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [64] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [71] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [78] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [85] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [92] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>   [99] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [106] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [113] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [120] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [127] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [134] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [141] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [148] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [155] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [162] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [169] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [176] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [183] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [190] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [197] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [204] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [211] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [218] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [225] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [232] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [239] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [246] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [253] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [260] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [267] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [274] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [281] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [288] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [295] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [302] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [309] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [316] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [323] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [330] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [337] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [344] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [351] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [358] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [365] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [372] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [379] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [386] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [393] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [400] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [407] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [414] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [421] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [428] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [435] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [442] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [449] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [456] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [463] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [470] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [477] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [484] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [491] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [498] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [505] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [512] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [519] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [526] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [533] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [540] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [547] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [554] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [561] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [568] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [575] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [582] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [589] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [596] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [603] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [610] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [617] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [624] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [631] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [638] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [645] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [652] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [659] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [666] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [673] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [680] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [687] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [694] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [701] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [708] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [715] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [722] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [729] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [736] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [743] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [750] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [757] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [764] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [771] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [778] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [785] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [792] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [799] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [806] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [813] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [820] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [827] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [834] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [841] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [848] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [855] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [862] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [869] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [876] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [883] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [890] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [897] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [904] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [911] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [918] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [925] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [932] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [939] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [946] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [953] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [960] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [967] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [974] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [981] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [988] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#>  [995] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1002] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1009] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1016] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1023] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1030] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1037] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1044] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1051] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1058] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1065] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1072] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1079] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1086] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1093] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1100] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1107] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1114] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1121] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1128] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1135] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1142] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1149] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1156] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1163] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1170] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1177] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1184] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1191] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1198] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1205] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1212] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1219] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1226] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1233] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1240] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1247] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1254] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1261] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1268] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1275] 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364 0.9038364
#> [1282] 0.9038364 0.9038364
osi_c_magnesium_uk(B_LU = c(265,1019),A_SOM_LOI = c(3,5),A_MG_AN = c(35,55))
#> Warning: Item 1 has 2 rows but longest item has 1283; recycled with remainder.
#>    [1] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>    [8] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [15] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [22] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [29] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [36] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [43] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [50] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [57] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [64] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [71] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [78] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [85] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [92] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>   [99] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [106] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [113] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [120] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [127] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [134] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [141] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [148] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [155] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [162] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [169] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [176] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [183] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [190] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [197] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [204] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [211] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [218] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [225] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [232] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [239] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [246] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [253] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [260] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [267] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [274] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [281] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [288] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [295] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [302] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [309] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [316] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [323] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [330] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [337] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [344] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [351] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [358] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [365] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [372] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [379] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [386] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [393] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [400] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [407] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [414] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [421] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [428] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [435] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [442] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [449] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [456] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [463] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [470] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [477] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [484] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [491] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [498] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [505] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [512] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [519] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [526] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [533] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [540] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [547] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [554] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [561] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [568] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [575] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [582] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [589] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [596] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [603] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [610] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [617] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [624] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [631] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390
#>  [638] 0.7304390 0.7304390 0.7304390 0.7304390 0.7304390 0.9133458 0.9133458
#>  [645] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [652] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [659] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [666] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [673] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [680] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [687] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [694] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [701] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [708] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [715] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [722] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [729] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [736] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [743] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [750] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [757] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [764] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [771] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [778] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [785] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [792] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [799] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [806] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [813] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [820] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [827] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [834] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [841] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [848] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [855] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [862] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [869] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [876] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [883] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [890] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [897] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [904] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [911] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [918] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [925] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [932] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [939] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [946] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [953] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [960] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [967] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [974] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [981] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [988] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#>  [995] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1002] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1009] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1016] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1023] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1030] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1037] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1044] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1051] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1058] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1065] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1072] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1079] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1086] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1093] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1100] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1107] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1114] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1121] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1128] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1135] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1142] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1149] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1156] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1163] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1170] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1177] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1184] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1191] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1198] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1205] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1212] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1219] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1226] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1233] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1240] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1247] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1254] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1261] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1268] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1275] 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458 0.9133458
#> [1282] 0.9133458 0.9133458